MongoDB prep js file wasn't deleting databases

+++ prep_scripts/dev_prepmongo.js
@@ -1,11 +1,15 @@
+const defaultDatabases = ["admin", "config", "local"]
+
 // Create mongo connection
 let mongo = new Mongo();
 
 // Remove old databases
 let db = mongo.getDB("test");
-let oldDbs = db.adminCommand('listDatabases')
+let oldDbs = db.adminCommand('listDatabases').databases;
 for (let i = 0; i < oldDbs.length; i++) {
-    oldDbs[i].dropDatabase();
+    // Make sure we don't delete default databases
+    if(!defaultDatabases.includes(oldDbs[i].name))
+        oldDbs[i].dropDatabase();
 }
 
 // Insert test records
